Remove cli functionality from dogecoind
As it says on the tin. It was deprecated in Bitcoin version 0.9, and
at some point it should be removed.

Removes the dependency of dogecoind on libbitcoin-cli.a. Move
some functions that used to be shared but are now only used in
dogecoin-cli.cpp to that file.

After this change, an error is printed (and exit code 1 is returned)
when the user tries to send RPC commands using dogecoind.
