+++ src/rpcclient.h
@@ -10,16 +10,6 @@
 #include "json/json_spirit_utils.h"
 #include "json/json_spirit_writer_template.h"
 
-int CommandLineRPC(int argc, char *argv[]);
-
 json_spirit::Array RPCConvertValues(const std::string &strMethod, const std::vector<std::string> &strParams);
-
-/** Show help message for dogecoin-cli.
- * The mainProgram argument is used to determine whether to show this message as main program
- * (and include some common options) or as sub-header of another help message.
- *
- * @note the argument can be removed once dogecoin-cli functionality is removed from bitcoind
- */
-std::string HelpMessageCli(bool mainProgram);
 
 #endif
